Create api.js file
