+++ src/index.html
@@ -16,10 +16,6 @@
         <a id="refresh" class="btn">Refresh</a>
       </div>
       <ul id="list" class="list">
-        <li class="item">name : 1</li>
-        <li class="item">name : 2</li>
-        <li class="item">name : 3</li>
-        <li class="item">name : 4</li>
       </ul>
     </div>
     <div class="content-holder d-flex-col">
@@ -29,13 +25,12 @@
       <div class="d-flex-col gap-2rem">
         <input id="user" type="text" class="input" placeholder="Your name">
         <input id="score" type="text" class="input" placeholder="Your score">
+        <p id="error" class="error-msg"></p>
         <a id="submit" class="btn submit-btn">Submit</a>
       </div>
     </div>
   </div>
 
-
-  <script src="main.js"></script>
 </body>
 
 </html>
