+++ src/style.css
@@ -13,7 +13,6 @@ html {
 
 .title {
   font-size: 4.5rem;
-  color: var(--color-primary);
   text-align: center;
   padding: 3rem 0;
 }
@@ -67,7 +66,7 @@ html {
 }
 
 .list {
-  border: 2px solid var(--color-primary);
+  border: 2px solid black;
   list-style-type: none;
   padding: 0;
 }
@@ -78,7 +77,7 @@ html {
 }
 
 .item:nth-child(odd) {
-  background-color: rgb(213, 227, 253);
+  /* background-color: rgb(213, 227, 253); */
 }
 
 .input {
@@ -91,7 +90,7 @@ html {
 
 .input:focus {
   outline: 0;
-  border-bottom: 2px solid var(--color-primary);
+  /* border-bottom: 2px solid var(--color-primary); */
 }
 
 .btn {
@@ -99,15 +98,20 @@ html {
   cursor: pointer;
   border: 0;
   background-color: white;
-  color: var(--color-primary);
+  /* color: var(--color-primary); */
   width: fit-content;
   transition: all 0.3s ease;
 }
 
 .btn:hover {
-  transform: scale(1.1);
-  color: var(--color-secondary);
-  border-bottom: 1px solid var(--color-secondary);
+  /* transform: scale(1.1); */
+  /* color: var(--color-secondary); */
+  /* border-bottom: 1px solid var(--color-secondary); */
+}
+
+.error-msg {
+  color: red;
+  font-size: 1.5rem;
 }
 
 .submit-btn {
